Encapsulating insert test documents logic
